Moved controllers folder from project root to inside API.
