+++ api/routes/routes.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router();
-const apiController = require('../../controllers/apicontroller');
+const apiController = require('../controllers/apicontroller');
 const path = require('path');
 require('dotenv').config({ path: path.resolve(__dirname, '../.env') });
